pyproject.toml
[project]
name = "pdf-2-anki"
version = "0.1.0"
description = "PDF to Anki flashcard generator using LangChain and Ollama"
requires-python = ">=3.11"
dependencies = [
    "langchain-ollama",
    "pymupdf",
    "pillow",
    "flask>=3.1.3",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
    "apyanki>=0.19.0",
]

[tool.uv]
dev-dependencies = []
